pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = ""
authors = [
    {name = "romadev",email = "romacreator24@gmail.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.30"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.29"
alembic = "^1.13"
python-jose = {extras = ["cryptography"], version = "^3.3"}
passlib = {extras = ["bcrypt"], version = "^1.7"}
bcrypt = "3.2.0"
pydantic = {extras = ["email"], version = "^2.7"}
pydantic-settings = "^2.3.4"
pytest-asyncio = "^0.23.7"
httpx = "^0.28.1"
redis = {extras = ["hiredis"], version = "^4.6.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
fastapi-limiter = "^0.1.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
aiosqlite = "^0.20.0"
fakeredis = "^2.30.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
